move rtl.js to common folder, update mapbox version
