+++ static/common/rtl.js
